requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Mail==0.9.1
Flask-Session==0.3.2
Werkzeug==2.0.1
email-validator==1.1.3
dnspython==2.1.0
